fet: create debug command unit
